basic scene exporter
Former-commit-id: a96a46a44b4bb60258a64d79c799b247437a0e81
Former-commit-id: 2b0728e12c481d84fd856db62d96750ec35ae4bf

+++ src/core/base64.cpp
@@ -13,7 +13,7 @@ namespace phi
         return (isalnum(c) || (c == '+') || (c == '/'));
     }
 
-    string base64::encode(byte const* buf, uint bufLen)
+    string base64::encode(byte const* buf, size_t bufLen)
     {
         string ret;
         int i = 0;
